refactor(illustration-exhibit-list): dont use pipe to filter
